feat: clean legacy question code
