--- apps/bilan-carbone/src/components/dynamic-form/inputFields/utils.ts
@@ -1,1 +1,0 @@
-export const formatOption = (option: string) => option.replace(/^\d*-/, '')
