--- apps/bilan-carbone/src/components/dynamic-form/types/questionTypes.ts
@@ -1,43 +1,0 @@
-import { QuestionType } from '@abc-transitionbascarbone/db-common/enums'
-
-export { QuestionType, SubPost } from '@abc-transitionbascarbone/db-common/enums'
-
-export enum TextInputFormat {
-  Text = 'Text',
-  Phone = 'Phone',
-}
-
-export enum NumberInputFormat {
-  Number = 'Number',
-  PostalCode = 'PostalCode',
-}
-
-export enum FieldType {
-  TEXT = 'TEXT',
-  NUMBER = 'NUMBER',
-  SELECT = 'SELECT',
-  SELECT_FE = 'SELECT_FE',
-  DATE = 'DATE',
-  YEAR = 'YEAR',
-  TITLE = 'TITLE',
-  QCM = 'QCM',
-  QCU = 'QCU',
-  TABLE = 'TABLE',
-  CUSTOM = 'CUSTOM',
-}
-
-export const QUESTION_TYPE_FIELD_MAPPING: Record<QuestionType, FieldType> = {
-  [QuestionType.TEXT]: FieldType.TEXT,
-  [QuestionType.NUMBER]: FieldType.NUMBER,
-  [QuestionType.SELECT]: FieldType.SELECT,
-  [QuestionType.QCM]: FieldType.QCM,
-  [QuestionType.QCU]: FieldType.QCU,
-  [QuestionType.TABLE]: FieldType.TABLE,
-  [QuestionType.POSTAL_CODE]: FieldType.NUMBER,
-  [QuestionType.DATE]: FieldType.DATE,
-  [QuestionType.RANGE]: FieldType.TEXT, // TODO: change when range selector is implemented
-  [QuestionType.PHONE]: FieldType.TEXT,
-  [QuestionType.SELECT_FE]: FieldType.SELECT_FE,
-  [QuestionType.TITLE]: FieldType.TITLE,
-  [QuestionType.CUSTOM]: FieldType.CUSTOM,
-}
