--- apps/bilan-carbone/src/components/modals/SiteDataChangeWarningModal.tsx
@@ -1,89 +1,0 @@
-'use client'
-
-import classNames from 'classnames'
-import { useTranslations } from 'next-intl'
-import Modal from './Modal'
-import styles from './SiteDataChangeWarningModal.module.css'
-
-interface Props {
-  isOpen: boolean
-  onClose: () => void
-  onConfirm: () => void
-  questionsBySubPost: Record<string, Array<{ id: string; label: string; idIntern: string; answer?: string }>>
-}
-
-const SiteDataChangeWarningModal = ({ isOpen, onClose, onConfirm, questionsBySubPost }: Props) => {
-  const t = useTranslations('study.perimeter.siteDataChangeModal')
-  const tPost = useTranslations('emissionFactors.post')
-
-  // Filter to only show subposts and questions that have answers
-  const filteredQuestionsBySubPost: Record<
-    string,
-    Array<{ id: string; label: string; idIntern: string; answer?: string }>
-  > = Object.fromEntries(
-    Object.entries(questionsBySubPost)
-      .map(([subPost, questions]) => [
-        subPost,
-        questions.filter((question) => question.answer && question.answer.trim() !== ''),
-      ])
-      .filter(([, questions]) => questions.length > 0),
-  )
-
-  const affectedSubPostsCount = Object.keys(filteredQuestionsBySubPost).length
-
-  return (
-    <Modal
-      open={isOpen}
-      onClose={onClose}
-      title={t('title')}
-      label="site-data-change-warning"
-      actions={[
-        {
-          actionType: 'button',
-          color: 'error',
-          children: t('cancel'),
-          onClick: onClose,
-        },
-        {
-          actionType: 'button',
-          children: t('proceed'),
-          onClick: onConfirm,
-        },
-      ]}
-    >
-      <div className={styles.modalBody}>
-        <p>{t('description')}</p>
-
-        {affectedSubPostsCount > 0 && (
-          <div>
-            <p className="mb1">{t('affectedCalculations')}</p>
-
-            {Object.keys(filteredQuestionsBySubPost).length > 0 && (
-              <div className={`my2 p1 ${styles.warningSection}`}>
-                <ul>
-                  {Object.entries(filteredQuestionsBySubPost).map(([subPost, questions]) => (
-                    <li key={subPost}>
-                      <strong>{tPost(subPost)}</strong>
-                      <ul className="mt-2 ml1 pl1">
-                        {questions.map((question) => (
-                          <li key={question.id} className="mb-2">
-                            {question.label}
-                            <div className={classNames(styles.questionAnswer, 'mb1 mt-2 p-2')}>{question.answer}</div>
-                          </li>
-                        ))}
-                      </ul>
-                    </li>
-                  ))}
-                </ul>
-              </div>
-            )}
-          </div>
-        )}
-
-        <p>{t('confirmation')}</p>
-      </div>
-    </Modal>
-  )
-}
-
-export default SiteDataChangeWarningModal
