--- apps/bilan-carbone/src/constants/questions.ts
@@ -1,46 +1,0 @@
-// Fonctionnement
-export const RENOVATION_SELECTION_QUESTION_ID = 'de-quel-type-de-renovation-sagit-il'
-export const INCREASE_SURFACE_QUESTION_ID =
-  'dans-le-cas-dun-agrandissement-quelle-est-la-surface-supplementaire-ajoutee'
-export const RENOVATION_QUESTION_ID = 'quel-est-le-montant-des-depenses-liees-a-ces-travaux-de-renovation'
-export const SERVICES_QUESTION_ID = 'quel-montant-avez-vous-depense-en-services'
-export const SHARED_ACTIVITY_QUESTION_ID = 'le-batiment-est-il-partage-avec-une-autre-activite'
-export const RESEAU_CHALLEUR_QUESTION_ID = 'reseaux-urbains-chaleur'
-export const RESEAU_FROID_QUESTION_ID = 'reseaux-urbains-chaleurfroid'
-export const GENERATOR_QUESTION_ID = 'le-cinema-dispose-t-il-dun-ou-plusieurs-groupes-electrogenes'
-
-// Mobilité spectateurs
-export const MOBILIY_SURVEY_QUESTION_ID = 'avez-vous-deja-realise-une-enquete-mobilite-spectateurs'
-export const SPECTATOR_SHORT_DISTANCE_DETAILS_QUESTION_ID =
-  '10-quelles-sont-les-distances-parcourues-au-total-sur-lannee-pour-chacun-des-modes-de-transport-suivants'
-export const MOBILITY_DOWNLOAD_MODEL_QUESTION_ID =
-  'question-en-cours-de-developpement-le-template-pourra-etre-telecharge-prochainement-si-vous-souhaitez-realiser-une-enquete-mobilite-spectateur-vous-pouvez-ici-telecharger-un-modele-denquete-qui-vous-permettra-de-remplir-dici-quelques-semaines-les-informations-demandees'
-export const SHORT_DISTANCE_QUESTION_ID =
-  'si-vous-souhaitez-vous-identifier-a-des-profils-de-cinema-comparable-de-quel-type-de-cinema-votre-etablissement-se-rapproche-le-plus'
-export const LONG_DISTANCE_QUESTION_ID =
-  'vos-spectateurs-sont-ils-majoritairement-des-habitants-locaux-cest-a-dire-residant-a-lannee-dans-les-environs-ou-attirez-vous-aussi-une-part-non-negligeable-de-spectateurs-de-passage-dans-la-region-touristes-notamment'
-
-// Confiseries et boissons
-export const CONFECTIONERY_SELECT_QUESTION_ID = 'vendez-vous-des-boissons-et-des-confiseries'
-export const CONFECTIONERY_QUESTION_ID = 'selon-vous-en-moyenne-un-spectateur-achete'
-
-// Salles et cabines
-export const ROOM_QUESTION_ID = 'decrivez-les-differentes-salles-du-cinema'
-
-// Tournees avant première
-export const MOVIE_TEAM_QUESTION_ID = 'combien-dequipes-de-films-avez-vous-recu-en'
-export const MOVIE_TEAM_MEAL_QUESTION_ID = 'combien-dequipes-de-repas'
-export const MOVIE_DEMAT_QUESTION_ID = 'combien-de-films-recevez-vous-en-dematerialise-par-an'
-export const MOVIE_DCP_QUESTION_ID = 'combien-de-films-recevez-vous-sur-dcp-physique-par-an'
-
-// Billetterie et communication
-export const NEWSLETTER_QUESTION_ID = 'combien-de-newsletters-ont-ete-envoyees'
-export const NEWSLETTER_RECEIVER_COUNT_QUESTION_ID = 'en-moyenne-a-combien-de-personnes-sont-adressees-ces-newsletters'
-
-// Dêchets
-export const XENON_LAMPS_QUESTION_ID = 'quelle-quantite-de-lampes-xenon-jetez-vous-par-an'
-
-export const CLIMATISATION_QUESTION_ID = 'votre-cinema-est-il-equipe-de-la-climatisation'
-
-// Calculation constants
-export const LONG_DISTANCE_APPLIED_PERCENTAGE = 0.03
