+++ apps/bilan-carbone/src/css/index.css
@@ -4,3 +4,4 @@
 @import './globals.css';
 @import './flex.css';
 @import './var.css';
+@import './inputUnit.css';
