+++ apps/bilan-carbone/src/css/inputUnit.css
@@ -8,6 +8,6 @@
   top: 1.75rem; /* Align with MUI TextField input baseline */
   right: 2rem;
   transform: translateY(-50%);
-  pointer-events: none; /* Prevent interfering with input clicks */
+  pointer-events: none;
   z-index: 1;
 }
