+++ apps/bilan-carbone/src/environments/simplified/study/infography/SimplifiedPostInfography.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Post, SimplifiedPost } from '@/services/posts'
-import { QuestionStats } from '@/services/serverFunctions/question'
+import { QuestionStats } from '@/services/publicodes/questionProgress'
 import { ResultsByPost } from '@/types/study.types'
 import { SubPost } from '@abc-transitionbascarbone/db-common/enums'
 import CheckCircleOutlineSharp from '@mui/icons-material/CheckCircleOutlineSharp'
