--- apps/bilan-carbone/src/hooks/useConditionalVisibility.ts
@@ -1,93 +1,0 @@
-import { emissionFactorMap } from '@/constants/emissionFactorMap'
-import { cleanupHiddenQuestion } from '@/services/serverFunctions/question'
-import type { Question } from '@abc-transitionbascarbone/db-common/types'
-import { useEffect, useMemo, useRef } from 'react'
-import { UseFormSetValue, UseFormWatch } from 'react-hook-form'
-
-export const useConditionalVisibility = (
-  questions: Question[],
-  watch: UseFormWatch<Record<string, unknown>>,
-  studySiteId?: string,
-  setValue?: UseFormSetValue<Record<string, unknown>>,
-) => {
-  const watchedValues = watch()
-  const previousVisibleQuestionsRef = useRef<Set<string>>(new Set())
-
-  const visibleQuestions = useMemo(() => {
-    const conditionallyVisibleQuestions = questions.filter((question) => {
-      const emissionInfo = emissionFactorMap[question.idIntern]
-
-      if (!emissionInfo?.conditionalRules || emissionInfo.conditionalRules.length === 0) {
-        return true
-      }
-
-      return emissionInfo.conditionalRules.every((rule) => {
-        const parentValue = watchedValues[rule.idIntern]
-
-        if (parentValue === undefined || parentValue === null || parentValue === '') {
-          return false
-        }
-
-        if (typeof parentValue === 'string' && parentValue.startsWith('[')) {
-          const parsedValue = JSON.parse(parentValue)
-          if (Array.isArray(parsedValue)) {
-            return rule.expectedAnswers.some((expectedAnswer) => parsedValue.includes(expectedAnswer))
-          }
-        }
-
-        return rule.expectedAnswers.includes(String(parentValue))
-      })
-    })
-
-    // Hide children when parent table questions are hidden (10-)
-    const hiddenParentTables = questions
-      .filter((q) => !conditionallyVisibleQuestions.some((visible) => visible.idIntern === q.idIntern))
-      .filter((q) => q.idIntern.startsWith('10-'))
-
-    return conditionallyVisibleQuestions.filter((question) => {
-      const isChildOfHiddenParent = hiddenParentTables.some((hiddenParent) => {
-        const parentBase = hiddenParent.idIntern.replace(/^10-/, '')
-        const questionBase = question.idIntern.replace(/^\d+-/, '')
-        return questionBase === parentBase && question.idIntern.match(/^\d+-/) && !question.idIntern.startsWith('10-')
-      })
-
-      return !isChildOfHiddenParent
-    })
-  }, [questions, watchedValues])
-
-  useEffect(() => {
-    if (!studySiteId) {
-      return
-    }
-
-    const currentVisibleQuestions = new Set(visibleQuestions.map((q) => q.idIntern))
-    const previousVisibleQuestions = previousVisibleQuestionsRef.current
-
-    const newlyHiddenQuestions = [...previousVisibleQuestions].filter(
-      (idIntern) => !currentVisibleQuestions.has(idIntern),
-    )
-
-    newlyHiddenQuestions.forEach(async (questionIdIntern) => {
-      try {
-        if (setValue) {
-          setValue(questionIdIntern, '', { shouldValidate: false, shouldDirty: false })
-
-          const allFormValues = watchedValues
-          Object.keys(allFormValues).forEach((fieldName) => {
-            if (fieldName.startsWith(`${questionIdIntern}-`)) {
-              setValue(fieldName, '', { shouldValidate: false, shouldDirty: false })
-            }
-          })
-        }
-
-        await cleanupHiddenQuestion(questionIdIntern, studySiteId)
-      } catch (error) {
-        console.error(`Failed to cleanup answer and emission sources for question ${questionIdIntern}:`, error)
-      }
-    })
-
-    previousVisibleQuestionsRef.current = currentVisibleQuestions
-  }, [visibleQuestions, studySiteId, setValue, watchedValues])
-
-  return visibleQuestions
-}
