--- apps/bilan-carbone/src/utils/question.ts
@@ -1,112 +1,0 @@
-import { FormValues } from '@/components/dynamic-form/types/formTypes'
-import { FieldType } from '@/components/dynamic-form/types/questionTypes'
-import { EmissionFactorInfo } from '@/constants/emissionFactorMap'
-import { ID_INTERN_PREFIX_REGEX } from '@/constants/utils'
-import type { Answer, Question } from '@abc-transitionbascarbone/db-common'
-import { QuestionType } from '@abc-transitionbascarbone/db-common/enums'
-import { JsonObject } from '@prisma/client/runtime/client'
-
-export const answersToFormValues = (questions: Question[], answers?: Answer[]): FormValues => {
-  const formValues: FormValues = {}
-
-  questions.forEach((question) => {
-    const answer = answers?.find((a) => a.questionId === question.id)
-
-    if (answer && answer.response) {
-      if (ID_INTERN_PREFIX_REGEX.test(question.idIntern) && question.type !== FieldType.TABLE) {
-        const keys = Object.keys(answer.response as JsonObject)
-        keys.forEach((key) => {
-          formValues[`${question.idIntern}-${key}`] = (answer.response as JsonObject)[key]
-        })
-      } else {
-        formValues[question.idIntern] = answer.response
-      }
-    } else {
-      formValues[question.idIntern] = null
-    }
-  })
-
-  return formValues
-}
-
-export const getQuestionLabel = (questionType: QuestionType, tFormat: (slug: string) => string) => {
-  switch (questionType) {
-    case QuestionType.POSTAL_CODE:
-      return tFormat('postalCode')
-    case QuestionType.PHONE:
-      return tFormat('phone')
-    default:
-      return ''
-  }
-}
-
-export const isQuestionRelatedToPrevious = (
-  currentQuestionInfo: EmissionFactorInfo,
-  previousQuestion: Question,
-): boolean => {
-  if (!currentQuestionInfo?.conditionalRules || currentQuestionInfo.conditionalRules.length === 0) {
-    return false
-  }
-
-  return currentQuestionInfo.conditionalRules.some((rule) => rule.idIntern === previousQuestion.idIntern)
-}
-
-export const areQuestionsLinked = (
-  currentQuestionInfo: EmissionFactorInfo,
-  previousQuestion: Question,
-  previousQuestionInfo: EmissionFactorInfo | null,
-): boolean => {
-  // Direct parent-child relationship (current question depends on previous question)
-  if (currentQuestionInfo && isQuestionRelatedToPrevious(currentQuestionInfo, previousQuestion)) {
-    return true
-  }
-
-  // Sibling relationship - when multiple questions depend on the same parent
-  if (
-    currentQuestionInfo?.conditionalRules &&
-    currentQuestionInfo.conditionalRules.length > 0 &&
-    previousQuestionInfo?.conditionalRules &&
-    previousQuestionInfo.conditionalRules.length > 0
-  ) {
-    const hasSharedParent = currentQuestionInfo.conditionalRules.some((currentRule) =>
-      previousQuestionInfo.conditionalRules?.some((previousRule) => currentRule.idIntern === previousRule.idIntern),
-    )
-    if (hasSharedParent) {
-      return true
-    }
-  }
-
-  return false
-}
-
-export const hasCompleteTableRow = (response: { rows: Array<{ data: Record<string, unknown> }> }): boolean => {
-  return response.rows.some((row) => {
-    const data = row.data
-    const fieldValues = Object.values(data)
-    return fieldValues.length > 0 && fieldValues.every((value) => value !== null && value !== undefined && value !== '')
-  })
-}
-
-export const findTableChildren = (
-  parentQuestion: { idIntern: string; type: string },
-  allQuestions: Array<{ idIntern: string; type: string }>,
-): Array<{ idIntern: string; type: string }> => {
-  const parentBase = parentQuestion.idIntern.replace(/^10-/, '')
-  return allQuestions.filter(
-    (q) =>
-      q.type !== 'TABLE' &&
-      q.idIntern !== parentQuestion.idIntern &&
-      q.idIntern.match(/^\d+-/) &&
-      !q.idIntern.startsWith('10-') &&
-      q.idIntern.replace(/^\d+-/, '') === parentBase,
-  )
-}
-
-export const shouldHideConditionalQuestion = (parentAnswer: string, expectedAnswers: string[]): boolean => {
-  if (typeof parentAnswer === 'string' && parentAnswer.startsWith('[')) {
-    const parsedValue = JSON.parse(parentAnswer) as string[]
-    return !parsedValue.some((res) => expectedAnswers.includes(res))
-  } else {
-    return !expectedAnswers.includes(parentAnswer)
-  }
-}
