+++ apps/bilan-carbone/src/utils/tableInput.ts
@@ -1,4 +1,11 @@
-import { TableAnswer, TableRow } from '@/components/dynamic-form/types/formTypes'
+export interface TableRow {
+  id: string
+  data: Record<string, string>
+}
+
+export interface TableAnswer {
+  rows: TableRow[]
+}
 
 export const isTableAnswer = (response: unknown): response is TableAnswer => {
   return (
