--- packages/db-common/prisma/schema/bilan-carbone/question.prisma
@@ -1,72 +1,0 @@
-enum QuestionType {
-  QCM
-  QCU
-  SELECT
-  SELECT_FE
-  TABLE
-  POSTAL_CODE
-  DATE
-  NUMBER
-  RANGE
-  PHONE
-  TEXT
-  TITLE
-  CUSTOM
-
-  @@schema("bilan_carbone")
-}
-
-model Question {
-  id              String       @id @default(uuid())
-  idIntern        String       @unique @map("id_intern")
-  label           String
-  order           Int          @default(autoincrement())
-  subPost         SubPost      @map("sub_post")
-  type            QuestionType
-  unit            Unit?
-  possibleAnswers String[]     @map("possible_answers")
-  userAnswers     Answer[]
-  required        Boolean
-  createdAt       DateTime     @default(now()) @map("created_at")
-  updatedAt       DateTime     @default(now()) @updatedAt @map("updated_at")
-
-  @@map("questions")
-  @@schema("bilan_carbone")
-}
-
-model Answer {
-  id       String @id @default(uuid())
-  response Json
-
-  studySite   StudySite @relation(fields: [studySiteId], references: [id], onDelete: Cascade)
-  studySiteId String    @map("study_site_id")
-
-  question   Question @relation(fields: [questionId], references: [id])
-  questionId String   @map("question_id")
-
-  emissionSources AnswerEmissionSource[]
-
-  createdAt DateTime @default(now()) @map("created_at")
-  updatedAt DateTime @default(now()) @updatedAt @map("updated_at")
-
-  @@unique([questionId, studySiteId])
-  @@map("answers")
-  @@schema("bilan_carbone")
-}
-
-model AnswerEmissionSource {
-  id               String              @id @default(uuid())
-  answer           Answer              @relation(fields: [answerId], references: [id], onDelete: Cascade)
-  answerId         String              @map("answer_id")
-  emissionSource   StudyEmissionSource @relation(fields: [emissionSourceId], references: [id], onDelete: Cascade)
-  emissionSourceId String              @map("emission_source_id")
-
-  rowId        String? @map("row_id") // "row-1751892643065-t7uker0u0" - unique row identifier
-  emissionType String? @map("emission_type") // "meal", "transport", etc. - emission source type within row
-
-  createdAt DateTime @default(now()) @map("created_at")
-
-  @@unique([answerId, rowId, emissionType]) // One emission source per row per type
-  @@map("answer_emission_sources")
-  @@schema("bilan_carbone")
-}
